requirements_automation.txt
# Shared dependencies for automation playbooks and notebooks
PyYAML>=6.0

